+++ eodag/api/product/_assets.py
@@ -63,11 +63,9 @@ class AssetsDict(UserDict):
     def update(self, data: dict[str, Any]) -> None:  # type: ignore
         """Update assets"""
         super().update(data)
-        self.product.normalize()
 
     def __setitem__(self, key: str, value: dict[str, Any]) -> None:
         super().__setitem__(key, Asset(self.product, key, value))
-        self.product.normalize()
 
     def as_dict(self) -> dict[str, Any]:
         """Builds a representation of AssetsDict to enable its serialization
