+++ eodag/api/product/metadata_mapping.py
@@ -39,6 +39,7 @@ from shapely import wkt
 from shapely.geometry import MultiPolygon, Polygon
 from shapely.ops import transform
 
+from eodag.api.product._assets import Asset
 from eodag.types.queryables import Queryables
 from eodag.utils import (
     DEFAULT_PROJ,
@@ -62,6 +63,7 @@ if TYPE_CHECKING:
 
     from shapely.geometry.base import BaseGeometry
 
+    from eodag.api.product._product import EOProduct
     from eodag.config import PluginConfig
 
 logger = logging.getLogger("eodag.product.metadata_mapping")
@@ -1812,3 +1814,115 @@ def get_provider_queryable_key(
         return ""
     else:
         return eodag_key
+
+
+def normalize_bands(product: EOProduct):
+    """Normalize bands in product"""
+    if hasattr(product, "properties"):
+        product.properties = _stac_normalize_bands(product.properties)  # type: ignore
+    if hasattr(product, "assets"):
+        for key in product.assets:
+            product.assets[key] = _stac_normalize_bands(product.assets[key])  # type: ignore
+
+
+def _stac_normalize_bands(data: Union[dict, Asset]) -> Union[dict, Asset]:
+    """Normalize bands in product.properties or product.assets from STAC 1.0 to STAC 1.1"""
+    UNPREFIX_BAND_FIELDNAME = [
+        "name",
+        "description",
+        "data_type",
+        "nodata",
+        "unit",
+        "statistics",
+    ]
+    EXCLUDE_MOVE_TO_PARENT_BAND_FIELDNAME = ["name", "eo:common_name"]
+
+    # https://github.com/radiantearth/stac-spec/blob/v1.1.0/best-practices.md#bands
+    # Migrate band STAC 1.0 to 1.1
+    if isinstance(data, dict) or isinstance(data, Asset):
+
+        # Gather eo:band et raster:bands
+        bands: dict[str, Any] = {"eo:bands": [], "raster:bands": []}
+        hasData = False
+        for fieldname in bands:
+            if fieldname in data:
+                if isinstance(data[fieldname], list):
+                    bands[fieldname] = data[fieldname]
+                else:
+                    bands[fieldname] = [data[fieldname]]
+                hasData = True
+                del data[fieldname]
+
+        if hasData:
+            processed_bands = []
+
+            # migrate eo:bands > bands
+            if len(bands["eo:bands"]) > 0:
+                for item in bands["eo:bands"]:
+                    band = {}
+                    for key in item:
+                        if key in UNPREFIX_BAND_FIELDNAME:
+                            band[key] = item[key]
+                        else:
+                            band["eo:{}".format(key)] = item[key]
+                    processed_bands.append(band)
+
+            # migrate raster:bands > bands
+            if len(bands["raster:bands"]) > 0:
+                index = 0
+                for item in bands["raster:bands"]:
+                    band = processed_bands[index]
+                    for key in item:
+                        if key in UNPREFIX_BAND_FIELDNAME:
+                            band[key] = item[key]
+                        else:
+                            band["raster:{}".format(key)] = item[key]
+                    if index < len(processed_bands):
+                        processed_bands[index] = band
+                    else:
+                        processed_bands.append(band)
+                    index += 1
+
+            # When a property has same value for each band, have to be moved into parent scope
+            if len(processed_bands) > 0:
+                field_values: dict[str, Any] = {}
+
+                # Lists each distinct value for a field of the same name on each band
+                for band in processed_bands:
+                    for key in band:
+                        if key not in field_values:
+                            field_values[key] = []
+                        if band[key] not in field_values[key]:
+                            field_values[key].append(band[key])
+
+                    # Move band fields from asset to parent if all fields shared same value
+                    # (distincs values == 1)
+                    remove_band_fields = []
+                    for key in field_values:
+                        if (
+                            key not in EXCLUDE_MOVE_TO_PARENT_BAND_FIELDNAME
+                            and len(field_values[key]) == 1
+                        ):
+                            # All band have same value
+                            data[key] = field_values[key][0]
+                            # Tag field "to remove" from assets
+                            remove_band_fields.append(key)
+                del field_values
+
+            # Remove from assets field moved to parent
+            cleaned_bands = []
+            for band in processed_bands:
+                cleaned_band = {}
+                for key in band:
+                    if key not in remove_band_fields:
+                        cleaned_band[key] = band[key]
+                if len(list(cleaned_band.keys())) > 0:
+                    cleaned_bands.append(cleaned_band)
+            processed_bands = cleaned_bands
+            del cleaned_bands
+
+            # Remap éband" field if contains at least one value
+            if len(processed_bands) > 0:
+                data["bands"] = processed_bands
+
+    return data
