+++ eodag/plugins/search/cop_marine.py
@@ -34,6 +34,7 @@ from dateutil.utils import today
 
 from eodag import EOProduct
 from eodag.api.product import AssetsDict
+from eodag.api.product.metadata_mapping import normalize_bands
 from eodag.api.search_result import SearchResult
 from eodag.config import PluginConfig
 from eodag.plugins.search import PreparedSearch
@@ -294,7 +295,7 @@ class CopMarineSearch(StaticStacSearch):
         assets.update(additional_assets)
         product = EOProduct(self.provider, properties, collection=collection)
         product.assets = AssetsDict(product, assets)
-        product.normalize()
+        normalize_bands(product)
         return product
 
     def query(
